Add FTP, but not test on board

+++ 015_Common.h
@@ -46,7 +46,7 @@ typedef unsigned char bool;
 // For Common
 typedef unsigned char u8;
 typedef unsigned short u16;
-typedef unsigned int u32;
+typedef unsigned long u32;
 
 #endif
